Fix UX problem with research problem field (a blur caused the value to disappear)
The select package is not really suitable for this task. Since the behaviour we want is auto complete and not a select box. For a select box it makes sense that a new value has to be created explicitly. For now it works, but if there are any other UX problems, let's replace this package